% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indepssCTPR.R
\name{indepssCTPR}
\alias{indepssCTPR}
\title{Independent ssCTPR based on summary statistics (a.k.a. soft-thresholding)}
\usage{
indepssCTPR(
  coef,
  adj,
  lambda = exp(seq(log(0.001), log(0.1), length.out = 20)),
  lambda_ct,
  thr = 1e-04,
  maxiter = 10000,
  trace = 1
)
}
\arguments{
\item{coef}{vector/matrix of regression coefficients (\eqn{r})}

\item{adj}{vector of adjacency coefficients}

\item{lambda}{a vector of \eqn{\lambda}s}

\item{lambda_ct}{a vector of \eqn{\lambda_{ctp}}s}

\item{thr}{threshold to stop CD algorithm}

\item{maxiter}{the maximum number of iterations}

\item{trace}{controls the amount of output}
}
\value{
A list with the length equal to the number of lambda_ct, each element of the list has teh following elements
\item{lambda}{Same as \code{lambda} in input}
\item{beta}{A matrix of estimates of \eqn{\beta}}
}
\description{
Independent ssCTPR based on summary statistics (a.k.a. soft-thresholding)
}
\details{
A function to find the minimum of \eqn{\beta} in  
\deqn{f(\beta)=\beta'\beta - 2\beta'r + 2\lambda||\beta||_1 + \lambda_{ct}||\beta-\s{t}||^{2}}
where \eqn{r} is the vector of regression coefficients.
}
