% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge.ssCTPR.pipeline.R
\name{merge.ssCTPR.pipeline}
\alias{merge.ssCTPR.pipeline}
\title{Merge a list of ssCTPR.pipeline objects}
\usage{
\method{merge}{ssCTPR.pipeline}(...)
}
\arguments{
\item{...}{\code{ssCTPR.pipeline} objects, separated by commas}
}
\value{
A \code{ssCTPR.pipeline} object
}
\description{
Merge a list of ssCTPR.pipeline objects
}
