% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p2cor.R
\name{p2cor_ct}
\alias{p2cor_ct}
\title{Function to convert p-values to correlation via the t-statistic}
\usage{
p2cor_ct(p, n, traits, sign, min.n = max(n, 30)/10)
}
\arguments{
\item{p}{Matrix of p-values}

\item{n}{Sample size}

\item{traits}{The number of traits}

\item{sign}{A matrix giving the sign of the correlations (e.g. the log odds ratios)}

\item{min.n}{The minimum sample size to be considered a valid p-value}
}
\value{
A matrix of correlations
}
\description{
Function to convert p-values to correlation via the t-statistic
}
