% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{runElnet}
\alias{runElnet}
\title{Runs elnet with various parameters}
\usage{
runElnet(
  lambda,
  shrink,
  lambda_ct,
  fileName,
  r,
  adj,
  N,
  P,
  col_skip_pos,
  col_skip,
  keepbytes,
  keepoffset,
  thr,
  x,
  trace,
  maxiter,
  startvec,
  endvec
)
}
\arguments{
\item{shrink}{shrinkage parameter s}

\item{lambda_ct}{cross trait penalty parameter}

\item{fileName}{the file name of the reference panel}

\item{r}{a matrix of SNP-wise correlation with primary trait and/or beta estimates of secondary traits}

\item{adj}{a vector of SNP-wise adjacency coefficients between the primary and secondary traits}

\item{N}{number of individuals in the reference panel}

\item{P}{number of variants in reference file}

\item{col_skip_pos}{which variants should we skip}

\item{col_skip}{which variants should we skip}

\item{keepbytes}{required to read the PLINK file}

\item{keepoffset}{required to read the PLINK file}

\item{thr}{threshold}

\item{x}{a numeric vector of beta coefficients}

\item{trace}{if >1 verbose output}

\item{maxiter}{maximal number of iterations}

\item{startvec}{start position for each block}

\item{endvec}{end position for each block}

\item{lambda1}{a vector of lambdas}
}
\value{
a list of results
}
\description{
Runs elnet with various parameters
}
\keyword{internal}
