\name{ssCTPR-package}
\alias{ssCTPR-package}
\alias{ssCTPR}
\docType{package}
\title{\packageTitle{ssCTPR}}
\description{\packageDescription{ssCTPR}}
\section{Package Content}{\packageIndices{ssCTPR}}
\author{\packageAuthor{ssCTPR}}
\section{Maintainer}{\packageMaintainer{ssCTPR}}
\keyword{package}
