% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssCTPR.pipeline.R
\name{ssCTPR.pipeline}
\alias{ssCTPR.pipeline}
\title{Run ssCTPR with standard pipeline}
\usage{
ssCTPR.pipeline(
  cor,
  traits,
  adj = NULL,
  chr = NULL,
  pos = NULL,
  snp = NULL,
  A1 = NULL,
  A2 = NULL,
  ref.bfile = NULL,
  test.bfile = NULL,
  LDblocks = NULL,
  lambda = exp(seq(log(0.001), log(0.1), length.out = 20)),
  s = c(0.2, 0.5, 0.9, 1),
  lambda_ct = c(0, 0.06109, 0.1392, 0.24257),
  destandardize = F,
  trace = 1,
  exclude.ambiguous = TRUE,
  keep.ref = NULL,
  remove.ref = NULL,
  keep.test = NULL,
  remove.test = NULL,
  sample = NULL,
  cluster = NULL,
  max.ref.bfile.n = 20000,
  nomatch = FALSE,
  ...
)
}
\arguments{
\item{cor}{A matrix of SNP-wise correlation with primary trait, derived from \code{\link{p2cor}}, and beta of secondary traits if have any}

\item{traits}{The number of traits}

\item{adj}{A matrix of adjacency coefficients. NROW is the number of variants; NCOL is the number of secondary traits.}

\item{chr}{Together with \code{pos}, chromosome and position for \code{cor}}

\item{pos}{Together with \code{chr}, chromosome and position for \code{cor}}

\item{A1}{Alternative allele (effect allele) for \code{cor}}

\item{A2}{Reference allele for \code{cor} (One of \code{A1} or {A2} must be specified)}

\item{ref.bfile}{\code{bfile} (\href{https://www.cog-genomics.org/plink2/formats#bed}{PLINK binary format}, without .bed) for 
reference panel}

\item{test.bfile}{\code{bfile} for test dataset}

\item{LDblocks}{Either (1) one of "EUR.hg19", "AFR.hg19", "ASN.hg19", 
"EUR.hg38", "AFR.hg38", "ASN.hg38", to use blocks defined by Berisa and Pickrell (2015)
based on the 1000 Genome data, or (2) a vector to define LD blocks, 
or (3) a data.frame of regions in \href{https://www.ensembl.org/info/website/upload/bed.html}{bed format}}

\item{lambda}{to pass on to \code{\link{ssCTPR}}}

\item{s}{A vector of s}

\item{lambda_ct}{to pass on to \code{\link{ssCTPR}}}

\item{destandardize}{Should coefficients from \code{\link{ssCTPR}} be 
destandardized using test data standard deviations before being returned?}

\item{trace}{Controls the amount of output.}

\item{exclude.ambiguous}{Should ambiguous SNPs (C/G, A/T) be excluded?}

\item{keep.ref}{Participants to keep from the reference panel (see \code{\link{parseselect}})}

\item{remove.ref}{Participants to remove from the reference panel(see \code{\link{parseselect}})}

\item{keep.test}{Participants to keep from the testing dataset (see \code{\link{parseselect}})}

\item{remove.test}{Participants to remove from the testing dataset (see \code{\link{parseselect}})}

\item{sample}{Sample size of the random sample taken of ref.bfile}

\item{cluster}{A \code{cluster} object from the \code{parallel} package for parallel computing}

\item{max.ref.bfile.n}{The maximum sample size allowed in the reference panel}

\item{...}{parameters to pass to \code{\link{ssCTPR}}}
}
\value{
A \code{ssCTPR.pipeline} object with the following elements
\item{beta}{A list of ssCTPR coefficients: one list element for each \code{s}}
\item{test.extract}{A logical vector for the SNPs in \code{test.bfile} that are used in estimation.}
\item{also.in.refpanel}{A logical vector for the SNPs in \code{test.bfile} that are used in \code{ssCTPR}.}
\item{sumstats}{A \code{data.frame} of summary statistics used in estimation.}
\item{sd}{The standard deviation for the testing dataset}
\item{test.bfile}{The testing dataset}
\item{keep.test}{Sample to keep in the testing dataset}
\item{ref.bfile}{The reference panel dataset}
\item{keep.ref}{Sample to keep in the reference panel dataset}
\item{lambda, s, lambda_ct, keep.test, destandardized}{Information to pass on to \code{\link{validate.ssCTPR.pipeline}}}
\item{pgs}{A matrix of polygenic scores}
\item{destandardized}{Are the coefficients destandardized?}
\item{exclude.ambiguous}{Were ambiguous SNPs excluded?}
}
\description{
The easy way to run ssCTPR
}
\details{
To run \bold{ssCTPR} we assume as a minimum you have a vector of summary 
statistics in terms of SNP-wise correlations of primary trait, vector(s) of effect sizes of secondary traits and their positions (\code{chr}, 
\code{pos}), one of \code{A1} or \code{A2}, and a reference panel, specified 
either in \code{ref.bfile} or \code{test.bfile}. If only \code{test.bfile} is specified, 
we assume \code{test.bfile} is also the \code{ref.bfile}.  
If only \code{ref.bfile} is specified, only ssCTPR coefficients are returned, 
and polygenic scores are not calculated.

If SNPwise correlations are not available, they can be converted from 
p-values using the function \code{\link{p2cor}}. 

\code{ssCTPR.pipeline} only uses those SNPs that are consistently defined
by \code{chr}, \code{pos}, \code{A1} and \code{A2} and the 
\href{https://www.cog-genomics.org/plink2/formats#bim}{PLINK .bim} files
specified with \code{ref.bfile} and \code{test.bfile} for estimation. 
\code{\link{matchpos}} is used to achieve this, which allows for 
flipping of SNP alleles in their definitions. The \code{beta} matrix in the output contains
all SNPs that are common to the summary statistics and \code{test.bfile}. 
However, \bold{ssCTPR} with \code{s} < 1 is only run on SNPs that are 
common to all of \code{ref.bfile}, \code{test.bfile} and the 
summary statistics. \bold{The ssCTPR coefficients for \code{s} < 1 are imputed with 
results from \code{ssCTPR} with s = 1 (soft-thresholding) 
run on SNPs that are common to \code{test.bfile} and the summary stats 
but not to \code{ref.bfile}.} To select only SNPs that are common to all 
three datasets, one can use the \code{also.in.refpanel} logical vector in the
output. 

For \code{keep.ref}, \code{remove.ref}, \code{keep.test}, and \code{remove.test}, 
see the documentation for \code{keep} and \code{remove} in \code{\link{ssCTPR}} 
for details.
}
\note{
Berisa, T. & Pickrell, J. K. 
Approximately independent linkage disequilibrium blocks in human populations. 
Bioinformatics 32, 283-285 (2015).
}
\examples{
\dontrun{
 ### Read summary statistics file ###
 ss <- fread("./data/summarystats.txt")
 head(ss)
 
 ### Convert p-values to correlations, assuming a sample size of 60000 for the p-values ###
 cor <- p2cor(p = ss$P_val.Y1, n = 60000, sign=ss$BETA.Y1))
 cor <- cbind(cor,ss$BETA.Y2)  # summary statistics of secondary traits
 
 ### Run ssCTPR using standard pipeline ### 
 out <- ssCTPR.pipeline(cor=cor, traits=ncol(cor), lambda_ct = lambda_ct, 
                          chr=ss$Chr, pos=ss$Position, 
                          A1=ss$A1, A2=ss$A2,
                          ref.bfile=ref.bfile, test.bfile=test.bfile, 
                          LDblocks = "EUR.hg19")
}
}
